git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@1926 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/packets/scene/LightUpdateTransformMessage.h
@@ -47,7 +47,7 @@ which carries forward this exception.
 
 #include "engine/engine.h"
 
-#include "../../objects/scene/SceneObjectImplementation.h"
+#include "../../objects/scene/SceneObject.h"
 
 class LightUpdateTransformMessage : public StandaloneBaseMessage {
 public:
@@ -67,9 +67,9 @@ public:
 		insertByte(0); // unknown
 		
 		// add direction
-		insertByte(scno->getDirectionAngle());
+		insertByte((byte) scno->getDirectionAngle());
 		
-		/*cout << "Position Update [" << player->getObjectID() << "] (" 
+		/*System::out << "Position Update [" << player->getObjectID() << "] (" 
 			 << (int) (player->getPositionX()) << "," << (int) (player->getPositionZ()) << "," 
 			 << (int) (player->getPositionY()) << ") - Dir = " << (int) (player->getDirectionAngle()) << "\n";*/
 	}
@@ -90,9 +90,9 @@ public:
 		insertByte(0); // unknown
 
 		// add direction
-		insertByte(scno->getDirectionAngle());
+		insertByte((byte) scno->getDirectionAngle());
 
-		/*cout << "Position Update [" << player->getObjectID() << "] (" 
+		/*System::out << "Position Update [" << player->getObjectID() << "] (" 
 				 << (int) (player->getPositionX()) << "," << (int) (player->getPositionZ()) << "," 
 				 << (int) (player->getPositionY()) << ") - Dir = " << (int) (player->getDirectionAngle()) << "\n";*/
 	}
