+++ src/server/zone/packets/scene/LightUpdateTransformWithParentMessage.h
@@ -68,9 +68,9 @@ public:
 		insertByte(0); // unknown
 		
 		// add direction
-		insertByte(object->getDirectionAngle());
+		insertByte((byte) object->getDirectionAngle());
 		
-		/*cout << "Position Update [" << player->getObjectID() << "] (" 
+		/*System::out << "Position Update [" << player->getObjectID() << "] (" 
 			 << (int) (player->getPositionX()) << "," << (int) (player->getPositionZ()) << "," 
 			 << (int) (player->getPositionY()) << ") - Dir = " << (int) (player->getDirectionAngle()) << "\n";*/
 	}
